+++ packages/framework/resources/views/pages/documentation-search.blade.php
@@ -1,7 +1,7 @@
 @extends('hyde::layouts.docs')
 @section('content')
     <h1>Search the documentation site</h1>
-    <style>#searchMenuButton, .edit-page-link {
+    <style>#search-menu-button, .edit-page-link {
             display: none !important;
         }